03_Source_Code/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
fastapi-mail==1.4.1
python-dotenv==1.0.1
pydantic[email]==2.9.2
pydantic-settings==2.5.2
slowapi==0.1.9
